requirements.txt
# Core NLP & RAG
langchain==0.1.20
langchain-community==0.0.38
langchain-text-splitters==0.0.1
sentence-transformers==2.7.0
faiss-cpu==1.9.0.post1
transformers==4.40.0
torch==2.3.0

# PDF Processing
PyMuPDF==1.24.3
pdfplumber==0.11.0

# Voice Processing
SpeechRecognition==3.10.1
gTTS==2.5.1
pyaudio==0.2.14

# Language Detection & Translation
langdetect==1.0.9
deep-translator==1.11.4

# QR Code
qrcode[pil]==7.4.2

# UI
streamlit==1.35.0

# Utilities
numpy==1.26.4
Pillow==10.3.0

# Claude AI Chatbot (new)
requests==2.31.0
